msm: vidc: Set a valid rc if PTR_ERR is 0
In cases where IS_ERR_OR_NULL() was true, the driver was forming the
return via PTR_ERR.  However if the OR_NULL part happens to be true,
PTR_ERR(NULL) would be 0, which would generally mean success.  So, if
PTR_ERR returns 0, assign a valid error code to rc.

Change-Id: I363c69963040074e7b97d85914ea168b8eba3ab3
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>