msm: vidc: Align width & height to 16 prior to calculating the load
NUM_MBS_PER_SEC assumes that the input width and height is already
aligned to 16.  However, callers of this macro don't necessarily ensure
this.  This leads to bugs especially with resolutions such as 1080p for
which the height can be 1080 or 1088.  Hence, instead of making the
assumption, align both parameters to 16 (i.e. macroblock dimensions)
prior to calculating the load.

Change-Id: Ic386d5cdf55fc9549e37439f4586e20c85719e87
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>