video: msm: mdss_fb: Fix backlight kernel bug
First of all, it's useless to call pdata->set_backlight after
a call that does that coming from stock QC driver...
...then, why are we supposed to set the backlight to something
different than ZERO, when we are in DISPLAY OFF mode?

Also, remove an ifdef for specific panel, as pdata retrieval has
to be done even for normal operation (a piece of code after ours
is using pdata).

That was entirely wrong.